YET ANOTHER try at getting BPM marks lined up correctly.
I've completely replaced the calculations for downsampling from samples to downsamples, and then from downsamples to pixels. 

The new formulas are widget-width agnostic, i.e. the ultimate goal is to display 1 second of waveform data in 100 pixels of screen space, allowing 
for oversampling so that GL will antialias the waveform view.


Everyone who has seen issues with the bpm marks drifting, please let me know if this has issues.

Please tell me if the cue point drifts also. 

Drifting in this case is defined to be moving at a different speed than the waveform. 

The BPM marks are often not aligned on a beat, and often drift with respect to the real BPM of the song, that is not what I mean in this situation. I 
need to know if I am drawing the BPM marks (even though they can be incorrect) at the incorrect place.