Making default bpm 40 instead of 60 -- looks basically the same..
Changing wavesummary to allow a different read-stride than samples->downsample length. 

Adding a neat feature to beat marks for people to take a look at.. they turn white as they pass by the center point. (Will need to make this skin 
customizable, but it looks good on outline). It's fantastic for songs whose bpm starts at the beginning of the song, though if the bpm is off by a 
constant offset then it becomes really clear.