Doh, accidentally committed SoundSourceMP3 when I meant to commit SoundSourceOggVorbis.
Undoing the former and committing the latter. 

Adding potential fix for bug #329442. Need confirmation, as it is a Windows bug.

This should fix crashes on Windows related to loading oggs and reading ogg headers. 

The problem was the use of ov_open. 

The vorbisfile documentation says that ov_open can cause segfaults on Windows. I switched us to use ov_fopen, which is safe for all platforms. 

I added some comments and fixed some other minor bugs (not closing the file handle when ov_open fails), added some Q_ASSERTS to parallel those in 
SoundSourceMP3, and changed the brace formatting to be uniform throughout. (it was different for every other set of braces :) ).