Cleanups / Changes to structure of the Waveform code that I should have done last summer.
- Adding a generic base class for all rendered elements named RenderObject. This way we can easily add/subtract functionality from WaveformRenderer.
- Ripped out all Background image drawing code from WaveformRenderer and stuck it in WaveformRenderBackground
- Ripped out currently used signal drawing code and stuck it in WaveformRenderSignal
- Changed WaveformRenderBeat and WaveformRenderMark to inherit RenderObject
- Cosmetic changes to SoundSourceOggVorbis.

In the future, we can make WaveformRenderer simply iterate over a list of RenderObjects and draw them. We can either insert them in some order or provide z-ordering or something. This will be essential to allow things like Sean's scripts to hook up new cue points into the waveform viewer.