Changing EngineBuffer so that it issues updates to periodic-update control objects (like playposition, rate) and visual_playposition every iteration of EngineBuffer::process().
Previously, if the Reader was busy, these actions would not occur. This behavior
can cause hiccups in the GUI because if the Reader hangs on an error in the
SoundSource, then the Engine will not update these indicators while the Reader
is reading. Bug #342113 is an example of this -- the waveform viewer stalls on
certain files that have MP3 frame errors because libmad stalls for too long, the
Reader is locked, and EngineBuffer does not update the visual_playpos control
object.