Make the cue mode save properly. There were a number of bugs at work here. Fixes bug #282485.
A brief description of the problems:

EngineBufferCue was using the cue_simple ControlObject for both a pushbutton
which fires the cue_simple functionality AND as the store for the current cue
mode. I made cue_mode the CO for storing the current cue mode. DlgPrefControls
was not properly loading the cue mode from config on startup, which is why the
mode was not persisting in the GUI selection box across restarts. Furthermore,
the outline skin uses cue_simple instead of cue_default, which is why outline
only showed Simple functionality.