New "Mark Tracks as Played" item in file menu: * Loads my customized logfiles (as of this version) and looks for Track Played events * marks those tracks as played.
This is useful for running mixxx twice but preserving what was played the first time.  VERY HACKY.
FIXME: Needs a logfile, but of course this is overwritten on program load! So this is actually a bit
dumb.
TODO: if this were to load cue files it would be useful to people other than me.