Fix library controls so they do not ignore no-op sets.
The collapsing of CO and COT caused a subtle change in behavior which
has broken MIDI controller use of LibraryControl's controls. In Mixxx
<=1.11, ControlObject::sync() would deliver control updates from
ControlObjectThreads to its proxies regardless of whether the control
ignored the update due to it being a no-op.

Now our control updates obey ignore-no-ops so both the CO and its COTs
do not receive an update if ignore-no-ops is enabled. Since this is
the intended behavior of ignoring no-ops, we should fix the faulty
controls rather than mimicking <=1.11.x behavior.

In this case, the problem is fixed for LibraryControl controls by
using ControlPushButton instead of ControlObject, since
ControlPushButton does not ignore no-ops.