Remove zap_denormal code.
Extremely small values (~1e-38) are slow on Intel, so we had code
to round small values ("denormals") to zero.  But it was causing
waveform discontinuities and therefore pops and clipping with the
new normalized range of -1 to 1.  Furthermore, when we compile in
optimized mode GCC will zap denormals for us anyway, so this code
is unneeded.