Remove WBaseWidget disabled state. Instead, use QWidget enabled property.
As discussed in Bug #1180872, we should allow skin writers to connect controls
to the enabled state of a widget to enable/disable widgets. This disables input
events for widgets that are disabled or whose parents are disabled.

* Remove old OnOff connection type. I haven't seen a single use of it since I
  joined the project in 2008.
* Convert PropertyBinder into a ControlWidgetConnection.
* Update WDisplay to check the QWidget enabled property.

Currently only WDisplay supports rendering itself in a disabled state. I'll add
support to more widgets in the future.