Support mono outputs and inputs.
* Support selecting mono outputs and inputs in Sound Hardware preferences.
* Microphone can accept stereo or mono input now.
* Vinyl control (and consequently deck passthrough) does not support mono.

Tested with a stereo built-in mic on the Vinyl Control passthrough, auxilliary
passthrough, and microphone inputs.

I think PortAudio, CoreAudio, or my hardware has a bug when a single output
channel is enabled. The sound Mixxx emits to Channel 1 can be heard faintly in
Channel 2 even if we have only opened Channel 1 for output. If we have only
opened one channel for output there should be no possible way Mixxx could cause
audio to be heard in Channel 2 so this suggests it is not a Mixxx bug.

Fixes Bug #681689.