Add user-visible name and description property for controls.
Eventually we can use this to automatically create tooltips based on widget
connected controls instead of using the tooltip_id in the skin. Also, for
dynamically changing controls (such as the Effect*Slot in the features_effects
branch) this allows us to change the tooltip on the fly based on the effect
loaded to the slot.