Add support for knobs with a center-point NOTE_ON/NOTE_OFF message.
Some controllers (e.g. the VCI-100) emit a center point NOTE_ON/NOTE_OFF message
instead of an 0x40. Up until now Mixxx has mistakenly mapped this to a button.

Now, if multiple controls are present the CC 7-bit message handler can still
fire as long as there is only a single CC control in the stream. All non-CC
messages are ignored.

Ideally we could map the NOTE_ON/NOTE_OFF to the control's reset control but
this will take a refactoring of LearningUtil / DlgPrefControllerLearning so I
left it as future work.

Added a test case for this type of knob to LearningUtilsTest.