Fix play verification when a seek is queued. Make LateNight-Stacked to use start_play.
Previously when play verification occurred, it would fail
if the play position was at the end of the track.  But if
a seek has been queued, that should still be allowed.  This was
causing the start_play button to fail every time the
sampler had been allowed to play the whole sample.