coverdao - fixing bug with the database schema: making LOCATION not unique
Currently we are identifying covers by looking for md5hash and it must be unique (obvious reasons).
In cases when the user overwrite the cover file in the trackdir, it will have the same location but an different md5hash.
In this way, coverdao would try to store the new cover (new md5hash) but it would not be possible since it would have the same location as the old image... so, as the md5hash already is unique, we do not need to mantain unique locations as well...