Make cover in dlgtrackinfo work with cancel
We want that changes to the cover in dlgtrackinfo behave similar to
changing other values of the track. This required me to rework how
WCoverArtMenu works. WCoverArtMenu is now only a common menu that DOES
NOT CHANGE the database instead the widget implementing the menu does
this now when it sees fit. Another important change to make this work is
that 'CoverArtCache::searchImage' can now be called from any class
without the need use 'requestCover'. This is necessary to reload the
cover from the ID3tags/folder without updating the DB. This is ok in
this case because we only search for 1 cover and we will only block the
UI thread for a fraction of a second. I also think this is ok in this
special case because changing covers won't happen very often and a user
will to this one track at a time anyway, while the normal cover display
requires us to provides covers for several tracks simultaneously.