load image when we know they are in the id3tag
with this change the search function is only called if we really don't
know the cover for a track or it cannot be loaded anymore. At all other
times we call the load function. As a nice side effect this reduces the
DB calls for tracks collections where most tracks store the coverart in
the id3tags.