LateNight-Stacked: Make sure 4deck and mixer states save properly.
There are two COs used to record the skin state: the widget stack index,
and the toggle state of the button.  Those two values must be aligned,
so that the first page of the stack corresponds to the "off" state
of the toggle button, and the second page corresponds to the "on"
state.