add covers to directory relocation
To be sure that covers are also migrated correctly when a directory is
relocated we have to update the location of the cover as well. This
would actually work normally but not in cases where a custom cover is
set and the track contains a cover in the ID3TAG. Then we would use the
ID3TAG cover again.