Clean up test config files
To have a clean test environment we have to delete all config files that
may have been created. This is very important for database related tests
during development. If we change the sql schema and dao queries the
tests won't work with the old library. This will make it look like the
tests are failing while everything is fine.