Did some testing with insanely short loops -- 300 is a decent number.
Below about 1000 samples, the actual nature of the sound being looped
is totally lost, and the waves start folding over in such a way that
the loop just starts producing a sine-like tone.  As the loop length is
halved, the tone doubles in octave.  Even at ~500 samples, the tone is
pretty silly and high-pitched, but it is audible.  At the previous
minimum length of 30, the tone is just a squeal.  Not really interesting.