Master Sync: if only one playing sync deck, reset user offset and reinit master beat distance.
This should help prevent user offsets from accumulating over the course
of a set, which is really annoying and makes master sync feel bad.
There's a chance that the master beat distance could be
incorrectly tweaked, which would be user-visible as tracks that were in
sync suddenly offset by a small amount.  I will do some on-decks testing
to make sure this isn't happening.