First-pass refactor of AutoDJ to make it more general.
Remove all the references to myCO1 and myCO2 and replace with
QLists of COs.  Use a signal mapper so slots aren't duplicated.
Started factoring out common pieces of code.

There is still a lot to do but this is a good first step toward
helping AutoDJ deal with a 4-deck situation.

TESTED=tests still pass, and hand-tested basic autodj functionality:

TODO: make automated tests for:
1) Start autoDJ -- it should load tracks as necessary and start the
first deck with the crossfader in the right place
2) As deck 1 nears end, start deck 2, animate crossfader, and load
new track in deck 1
3) Same test but go from deck 2 to deck 1
4) Same as 2, but while crossfade is happening, stop deck 1.  Crossfader
should immediately go all the way to deck 2
5) same test, but go from deck 2 to deck 1