Wait until the first track load request succeeds before removing the track from the queue.
If the first track in the queue fails to load then the track that is
eventually loaded is not removed from the queue and the second track
loaded by AutoDJ will be a duplicate of the first track. Fix this by
waiting until the first track starts playing in the ADJ_ENABLE_P1LOADED
state to remove the track from the queue.