LateNight: Implement SizeAwareStack for lower half of UI.
Now the effects rack, samplers, and library will appear and disappear depending on the
size of the window.  Each state needs its own COs, so there is some
odd jumping around as the window size changes.  But most users don't
spend a lot of time resizing the window so it's not a dealbreaker.

TODO: need to hide the "effects" and "samplers" buttons when Big Library
mode is engaged