Bug #1021857 : Added a tooltip for Auto Recall Cue
Hey, I am a newbie, and i thought this (Bug #1021857) would be a great one to start with.
I am unable to compile mixxx as for now, for various reasons, and i am sorry for that, and hence i did not make any changes in tooltips.cpp  as of now, but i hope this fixes the issue. If not, please let me know on what else is to be done.