Auto DJ: Create a simple bandaid for the case when decks 3 and 4 are playing.
This is not pretty, but we already do something similar when decks 1 and 2
are both playing, and we are out of time to design a nice way to handle a
"4-deck" autodj.

This will eventually be replaced with something nicer!

Also fixed the autodj button not getting un-pushed when the enabling failed.