Fix newly-loaded tracks affecting playing master sync bpm.
If the rate slider for a track had been moved, a newly-loaded
track could sometimes cause the master sync bpm to change,
causing a sudden rate change in the playing decks.

The cause was a signal/slot reentrance in which loading the track
causes synccontrol to set the "bpm" control object, and
that would sometimes loop back around until enginesync interpretted that
as a slider change, and would change the master sync bpm.

The fix was to pass the newly-loaded track's bpm as a suggested
bpm to enginesync, and then it can choose whether to use it or
to override the suggestion and use an existing master bpm instead.

I tried to write a test for this but failed -- the bug is only triggered
when the rate slider update fails an ignorenoop equality comparison in
control.cpp due to floating point fuzziness.  This turned out
to be really hard to fake in the testing framework.  But there's no
reason the signal loop should occur now, nor should it regress
since I added a comment.

Existing tests pass.