Remove parenting of BeatGrid/BeatMap to its associated TrackInfoObject and move BeatGrid/BeatMap to the associated TrackInfoObject's thread more consistently.
Fixes Bug #1407124. If a BeatGrid/BeatMap that was parented to its
TIO (was loaded from the database) was deleted from the analyzer
thread (i.e. when a new beatgrid replaced the old one) then a Qt
assertion is triggered due to child-deletion events being sent to the
TIO from the analyzer thread. To solve this we don't parent the
BeatMap/BeatGrid to anything.