Update .travis.yml
Use clang and GCC to build and test mixxx. It is also a recommended practice for Travis to update the virtual machine on every run.
I also changed the scons options to the ones we use on the jenkins build server. It mainly concerns some plugins that are disabled in the default settings.