midi for light controller
A controller with many midi outputs. Beat, vu-meter, timecode,
deckchanges, bpm.  For better light and  be in sync to the musik. Works
fully automated to and with the current deck and has no problem with
pitch. Talked (mailed) about this with daniel, he said make a pull in
git. I think to force communication with light software is a good
feature for mixxx ;-) . (write wiki page is next step).