Sound Preferences: Move Query button above the fold.
Even on my high-rez screen, the Query button is way at the bottom below the input/output selectors,
and there's so much white space it's hard to even notice that there's a scroll bar on the side.  This
patch puts the query button and the hint above the giant selector list.