Create a framework for testing actual output of the enginebuffer.
After a call to ProcessBuffer(), tests can now compare the contents
of the output buffer against a golden data file representing expected
output.  This is useful for catching regressions in the engine code
as well as seeing how changes in the engine code affect audio output.

Included in this commit is a script called AudioPlot.py which has
been very useful to me in the past for analyzing audio data. It
requires matplotlib to be installed.