change analog knobs to control filters, with shift+right knob controlling cue/master mix
After talking to daschuer on lp:1458213, I realize that headphone gain
should not regularly be manipulated while DJing. So, remove that from
the mapping and make the knobs control filters by default. With shift,
the right knob still controls cue/master mix and turning it all the way
to the right toggles split cue mode.