fix LED ring for hotcue page selection
There seems to be a bug in the Tweaker firmware that does not allow
local control of LED rings for some rings but not others. Before, when I
had local control enabled for the low encoder with shift pressed, it
made the middle encoder behave strangely, kinda like it also had local
control enabled even though local control had been disabled. To get
around this, disable local control for all LED rings with shift pressed
and manually handle sending MIDI signals to LED rings.

Also, completely fill the appropriate half of the LED ring for the
relative controls of moving loops in loop mode with the middle encoder
and coarse skipping through track with shift with the low encoder. These
are reset back to center by timers after 1 second. Timers are disabled
when leaving that mode.