chmod a+x res/controllers/convertToXMLSchemaV1.php
Fixes:

W: non-executable-script /usr/share/mixxx/controllers/convertToXMLSchemaV1.php 0644L /usr/bin/env
This text file contains a shebang or is located in a path dedicated for executables, but lacks the executable bits and cannot thus be executed. If the file is meant to be an executable script, add the executable bits, otherwise remove the shebang or move the file elsewhere.