Renaming: "frame rate" -> "sampling rate"
Use the common term "sampling rate" from AudioSignal instead of "frame rate".
The term "frame rate" is still valid, but we should avoid different names for
the same concept.

I consciously avoided the term "sample rate", because that sounds like "number
of samples per second" which is wrong for signals with more than a single
channel and might be misleading!