Escape characters in mountpoint metadata sent to server.
For odd case that the mountpoint contains querystring characters.
Note that this is not truly supported by IceCast2 servers
out-of-the-box, as the pages construct admin pages and filenames with
the unescaped source.

Also, shout_set_mount does not detect the first slash in a mount point
if it is encoded, so remove the initial slash before doing any work on
the string.