Increase number of MIDI messages to buffer, fix bug #1527410
Chatty MIDI controllers like the Stanton SCS.1d (moving platter) can queue up a large number of messages if the system is for whatever reason slow to respond to them. If more than this number queue up, Mixxx stops responding to that controller. Experimental stress-testing showed 500 messages at once in the queue (during stupid things like application window switching, but still.)
This may not be the best fix, but it's quick and breaks nothing for the 2.0 release.