Add MixxxSettings as QSettings wrapper
Our config file lies in such a 'weird' place that the normal mechanisms
of telling QSettings about it don't work. I need to give it the exact
filename when construting the QSettings instance. Since the whole point
of this exercise is to get rid of the m_pConfig pointer and have a
settings object easily available anywhere in mixxx I need to create this
wrapper which saves the location of the config file at the start of
mixxx. This way MixxxSettings instances can be created anywhere in mixxx
with quick access to any setting.