SoundSourceProxy must be constructed from a TIO
Removed one of the SoundSourceProxy constructors that allowed to create
instances by providing a file name. Now instances can only be created
with a TIO pointer.

This is a prerequisite for reliably tracking all readers/writers of an
audio file in Mixxx. All file access must be initiated through
TrackInfoObject.