Move responsibility for adding tracks to the library to TrackDAO
This fix avoids a race condition that becomes relevant when implementing
caching of tracks. The cache will not updated properly when adding track
through a bypass where the track is allocated before it is added to
the library. The cache needs to be updated atomically after a track
has been added to the library with the track id.