Improve extraction of state from QDomNodes in widget classes.
Many widgets and common widget parsing code in LegacySkinParser used an
inefficient method of accessing state:

if (context.hasNode(node, "Foo")) {
  QDomElement foo = context.selectElement(node, "Foo");
}

This is inefficient because of the double lookup (which is O(n) in XML
children). This code is at the core of parsing the entire widget tree so
the inefficiencies add up -- we run code of this pattern many hundreds
of times for one skin.