Make ShortcutController, new owner of ShortcutChangeWatchers
All ShortcutChangeWatchers are currently stored in a QList in WMainMenuBar.
This is a problem, because we also need those watchers in other parts of
Mixxx. (i.e. in LegacySkinParser, in order to keep tooltips up to date).
Make a member variable m_pShortcutController in Mixxx, which will then
pass the shortcutcontroller to whatever object needs to know about shortcut
updates. Signals will be received by the ShortcutController.