Make Tooltipupdater, updating widget tooltips containing kbd shortcuts
The creation of tooltips is currently a very linear process, that find place
in LegacySkinParser::setupConnections(). We need to extract the tooltips setup
and make it more generic, so that it can be refreshed when a new kbd preset
is loaded. When a new keyboard preset is loaded, the TooltipUpdater is passed
a clone of that kbd preset via a signal and refreshes all widgets, passed in
by addWatcher().