Fix PushButtonTooltipWatcher
There is a currently a bug, displaying tooltip shortcuts
that are bound to a WBaseObject that is nor a WSliderComposed,
nor a WPushButton. Fix by tweaking PushButtonTooltipWatcher::udate
make correct use of setTooltipShortcut method, that was updated
in the last commit, but the implementation in PushButtonTooltipWatcher
was not updated yet.