Add TooltipUpdater support for multiple ConfigKeys
Widgets can have multiple connections to multiple configKeys. With
the current implementation, if there are two connections, and the
second one's configKey has no keyboard shortcut defined, than it
picks that one (empty string) and the tooltip shows any shortcut.
With support for multiple config keys that problem is gone and the
tooltip always displays the shortcuts.