Add WidgetTooltipWatcher factory method to WBaseWidget
The creation of WidgetTooltipWatchers currently takes place in
TooltipShortcutUpdater. This is a problem, because we have to
downcast the passed WBaseWidget so that we know if we need a
SliderTooltipWatcher, a PushButtonTooltipWatcher or a regular
WidgetTooltipWatcher. To get rid of this downcast, move the
creation to virtual WBaseWidget::createTooltipWatcher() which
will be overridden by WPushButton and WSliderComposed.