Remove kbd controller enumerator and make kbd preset loaded signal
We only have one instance of KeyboardController, and therefore it's not
necessary to have a KeyboardEnumerator, because it will always query
just one device; one instance of KeyboardController. That's why it's
better to just create the instance of KeyboardController directly in
ControllerManager. Also, a reference to that one instance has to be
stored as a member variable of ControllerManager. The current workaround
is to iterate through all devices and if possible, downcast to
KeyboardController. That's not neat. If KeyboardController is
instantiated directly in ControllerManager, there is no need anymore
to downcast.

Also. Currently some parts of Mixxx that need passed in a keyboard
controller preset are connected to KeyboardController::presetLoaded(),
which sends a clone of the preset as a QSharedPointer<ControllerPreset>.
This is a problem, because we then need to downcast this controller
preset to QSharedPointer<KeyboardControllerPreset> in order to work
with it. That's why we need to make KeyboardController emit a new
signal: keyboardControllerPresetLoaded, passing directly a
QSharedPointer<KeyboardControllerPreset>, so there is no need for
downcasting anymore.