Let FakeController::getPreset() default to Midi
Currently getPreset checks in getPreset which type of
preset it should return by checking the booleans: m_bHidPreset,
m_bKbdPreset and m_midiPreset. If one is true, it returns
that type of preset. We have to default to some type though,
to avoid "control reaches end of non-void function" warning.