Remove absolute window sizes
Currently we have absolute window sizes for the main frame and
the dialog window, so that we can center them to the screen.
However, the widgets are bigger or smaller, depending on the
platform, which make the UI seem ugly. That's why we need
to not specify any width or height and let pack() do it's job.