Use QString instead of ConfigValueKbd
Currently we have a problem with the Greek keyboard that we
can not distinguish the W key and the S key from each other,
because both keys return the same key code, QKeyEvent::key().

However, they do generate different unicode text, that's why
need to not base our comparisons on key codes, but rather on
the generated unicode text by those keys. This is true for
unmodified keys, which generate a valid unicode character,
but not for modified keys which are not guaranteed to do so.
That's why we need to do comparisons based on strings for
unmodified key sequences, but keep using QKeySequences for
modified sequences.