Give focus to first widget to load
Currently there is any widget that has focus when Mixxx starts. That's
why the keyboard does not work straight away, but it needs the user to
manually give focus to one of the widgets by clicking on it. This is a
real pain, because it's not clear if the keyboard just does not work,
if the Keyboard Controller is turned off in preferences, if it's not
receiving events because there's no focus, etc ... That's why we have
to give focus to the first widget to load, so that the keyboard works
straight away when opening Mixxx.