Fix midi note-off should be always considered a button release
Some controllers, like the M-Audio Xponent or keyboards with
after-touch, that send a non-zero note off velocity.  A note off should
always be considered a button release, regardless of its velocity value.

We duplicated every test involving a note-off to make sure we test both
for the `velocity == 0` and `velocity != 0` case -- given the subtle
structure of the code, it is easy to break either case.  Alternative for
conciseness we could just test the later case.