account for P32 sending values other than 127 and 1 for encoders
When the encoders are turned quickly, the P32 sends values a little
above 1 for right turns and a little below 127 for left turns. The
encoders have to be turned really fast for this to occur, so I don't
think it is very useful. This change treats all turns of the encoders as
a single step regardless of how fast it is turned.