Prevent double-free of MixxxMainWindow::m_pWidgetParent.
QMainWindow::setCentralWidget(nullptr) calls QObject::deleteLater on the current
central widget. If the Qt delete event is processed first then we would
double-free the widget when we delete it manually.

To protect against this, use the same trick we use for WMainMenuBar to ask Qt to
process the deleteLater event by using
QCoreApplication::sendPostedEvents(widget, QEvent::DeferredDelete) instead of
deleting the widget ourselves.