Fix the reverb effect The previous implementation of the reverb effect was suffering from serious artifacts in terms of feedback / distortion that the original LADSPA plugin doesn't suffer from: https://bugs.launchpad.net/mixxx/+bug/1647568
I updated the code base to use the latest version (0.9.24) from the original author (http://quitte.de/dsp/caps.html), and in doing this I disabled the unused code with preprocessor directives instead of deleting it, in order to make it easier to update in the future.

Next I tried to track down where the bug was. It turns out that the `SampleUtil::linearCrossfadeBuffers()` call was causing all the feedback problems, and it seems to be safe to remove this. In the future we should use ramping on the parameters.

This commit also adds the missing two controls for damping and blend, and remaps the meta knob from bandwidth to blend, which fixes the following bug:
https://bugs.launchpad.net/mixxx/+bug/1647775

Note that the meaning of "blend" has changed compared to the original source code, as we use it as a "send amount" instead of wet/dry. This has the advantage that the reverb tail is not suddenly cut off when the blend knob is turned down to zero.

In the PR (#1113) it was discussed that in the future we should probably implement wet/dry and send functions at the system level rather than forcing individual effects to implement it.