introduce loopauto_toggle and loopmanual_toggle COs
These will be used for intuitive control of loops with just 2 buttons in
skins and controllers.

loopauto_toggle: If beatloop_size > 1, press to set a beatloop of that
size. If beatloop_size <= 1, set a rolling beatloop while button is
held.

loopmanual_toggle: If loop is disabled, set loop in point when pressed
and loop out point when released. If loop is enabled, exit the loop on
press.